refactor: split authentication middleware into two middleware authenticate and authorize
